// src/app/core/interfaces/sales-order.ts
import { AccountingAttributes } from "./accounting-attributes";
import { CustomFields } from "./custom-fields";
import { LineItem } from "./line-item";

export interface SalesOrder {
    id?: number;
    salesOrderId?: number;
    createdDate?: Date | null | string;
    modifiedDate?: Date | null;
    createdBy?: number;
    processedBy?: number;
    isApproved?: boolean;
    reference?: string;
    memberId?: number;
    firstName?: string;
    lastName?: string;
    company?: string;
    email?: string;
    phone?: string;
    mobile?: string;
    fax?: string;
    deliveryFirstName?: string;
    deliveryLastName?: string;
    deliveryCompany?: string;
    deliveryAddress1?: string;
    deliveryAddress2?: string;
    deliveryCity?: string;
    deliveryState?: string;
    deliveryPostalCode?: string;
    deliveryCountry?: string;
    billingFirstName?: string;
    billingLastName?: string;
    billingCompany?: string;
    billingAddress1?: string;
    billingAddress2?: string;
    billingCity?: string;
    billingPostalCode?: string;
    billingState?: string;
    billingCountry?: string;
    branchId?: number;
    branchEmail?: string;
    projectName?: string;
    trackingCode?: string;
    internalComments?: string;
    productTotal?: number;
    freightTotal?: number;
    freightDescription?: string;
    surcharge?: number;
    surchargeDescription?: string;
    discountTotal?: number;
    discountDescription?: string;
    total?: number;
    currencyCode?: string;
    currencyRate?: number;
    currencySymbol?: string;
    taxStatus?: string;
    taxRate?: number;
    customFields?: CustomFields;
    isVoid?: boolean;
    accountingAttributes?: AccountingAttributes;
    memberEmail?: string;
    memberCostCenter?: string;
    memberAlternativeTaxRate?: string;
    costCenter?: string;
    alternativeTaxRate?: string;
    estimatedDeliveryDate?: Date | string | null;
    salesPersonId?: number;
    isInCarton?: boolean;
    salesPersonEmail?: string;
    paymentTerms?: string;
    customerOrderNo?: string;
    voucherCode?: string;
    deliveryInstructions?: string;
    cancellationDate?: Date | null;
    modifiedCOGSDate?: Date | null;
    status?: string;
    stage?: string;
    invoiceDate?: Date | null;
    invoiceNumber?: number;
    dispatchedDate?: Date | null;
    logisticsCarrier?: string;
    logisticsStatus?: number;
    ediStatus?: number;
    distributionBranchId?: number;
    departmentNumber?: string;
    storeLocationNumber?: string;
    distributionCenter?: string;
    lineItems?: LineItem[];
    orderLineItems?: LineItem[];
    isMatchedWithWoolWorth?: boolean;
    bannerOrderId?: number;
    itemsStr?: string;
    bannerItemsStr?: string;
    isSendEmail?: boolean;
    replyEmail?: string;
    isNotValidatePoNumber?: boolean;
    storeGuidId?: string;
    customPoStatusId?: number;
    customPoGuidId?: string;
    proofOfDeliveryURL?: string;
    isUploadedProofOfDelivery?: boolean;
    salesOrderWithSamePoNo?: number;
    isInvoiceSent?: boolean;
    invoiceSendDate?: Date;
    eta?: string;
    isRequestedToSend?: boolean;
    invoicePDFURL?: boolean;
    isInvoiceSentFromBot?: boolean;
    emailIssues?: string;
    acknowledgementDate?: Date;
    isAcknowledgement?: boolean;
    isSynced?: boolean;
    paidAmount?: number;
    paidPercentage?: number;
    invoiceEmail?: string;
    orderMode?: string;
}